Moving to smaller classes and interfaces
